Disable rostrings, force copying all strings
AVRs have separate address spaces for RAM and flash, so this won't work.
This is also the reason why the tiny ram patch won't work as-is (it uses
the same stext and etext approach).

A proper fix to this would allow the platforms to implement these
functions in their own way.